refactor(enterprise): SMI-5036 split oversized billing test files
- StripeClient.test.ts (724L) -> 4 files: customers/checkout/subscriptions/portal-invoices
- webhook-handlers.test.ts (684L) -> 3 files: helpers/subscriptions/invoices
- Extracted shared mocks to _helpers.ts (not .test.ts per CLAUDE.md SMI-1780)
- Removed 2 grandfather entries from scripts/check-file-length.ignore
- All resulting files <500 lines
- Test count unchanged (171 + 1 skip), all passing
